[FIX]: styling cleanup
