+++ src/features/offer/components/OfferNewXPCine/OfferNewXPCineBlock.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useMemo, useRef } from 'react'
-import { FlatList, StyleProp, View, ViewStyle } from 'react-native'
+import { FlatList, View } from 'react-native'
 import styled, { useTheme } from 'styled-components/native'
 
 import { OfferResponseV2 } from 'api/gen'
@@ -59,29 +59,24 @@ export function OfferNewXPCineBlock({
     }
   }, [flatListRef, movieScreeningDates, setSelectedDate])
 
-  const flatListPadding: StyleProp<ViewStyle> = theme.isDesktopViewport
-    ? {
-        paddingHorizontal: 0,
-      }
-    : undefined
-
   return (
     <Container testID="offer-new-xp-cine-block">
-      <CineBlockContainer>
+      <TitleContainer>
         <TypoDS.Title3 {...getHeadingAttrs(2)}>{title}</TypoDS.Title3>
-      </CineBlockContainer>
+      </TitleContainer>
 
       <Spacer.Column numberOfSpaces={4} />
 
-      <MovieCalendar
-        dates={movieScreeningDates}
-        selectedDate={selectedDate}
-        onTabChange={setSelectedDate}
-        flatListRef={flatListRef}
-        flatListPadding={flatListPadding}
-      />
+      <MovieCalendarContainer>
+        <MovieCalendar
+          dates={movieScreeningDates}
+          selectedDate={selectedDate}
+          onTabChange={setSelectedDate}
+          flatListRef={flatListRef}
+        />
+      </MovieCalendarContainer>
 
-      <CineBlockContainer>
+      <View>
         <CineBlock
           offer={offer}
           distance={distance}
@@ -89,7 +84,9 @@ export function OfferNewXPCineBlock({
           CTAOfferModal={CTAOfferModal}
           eventCardData={eventCardData}
         />
-      </CineBlockContainer>
+        <Spacer.Column numberOfSpaces={theme.isDesktopViewport ? 6 : 4} />
+        <Divider />
+      </View>
     </Container>
   )
 }
@@ -98,6 +95,15 @@ const Container = styled(View)({
   marginVertical: 0,
 })
 
-const CineBlockContainer = styled(View)(({ theme }) => ({
-  marginHorizontal: theme.isDesktopViewport ? undefined : getSpacing(6),
+const MovieCalendarContainer = styled(View)(({ theme }) => ({
+  marginRight: theme.isDesktopViewport ? -getSpacing(16) : 0, // cancels padding of the parent container
+}))
+
+const TitleContainer = styled(View)(({ theme }) => ({
+  marginHorizontal: theme.isDesktopViewport ? undefined : theme.contentPage.marginHorizontal,
+}))
+
+const Divider = styled.View(({ theme }) => ({
+  height: 1,
+  backgroundColor: theme.colors.greyMedium,
 }))
